+++ app/controllers/staff/activities_controller.rb
@@ -20,7 +20,7 @@ class Staff::ActivitiesController < Staff::BaseController
     authorize @activity
 
     @activity.hierarchy = hierarchy
-    @activity.set_hierarchy_defaults
+    @activity.wizard_status = "identifier"
     @activity.save(validate: false)
 
     redirect_to url_for([@activity.hierarchy, @activity, :steps])