+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_12_13_123607) do
+ActiveRecord::Schema.define(version: 2019_12_17_160336) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -35,6 +35,7 @@ ActiveRecord::Schema.define(version: 2019_12_13_123607) do
     t.datetime "updated_at", precision: 6, null: false
     t.string "hierarchy_type"
     t.uuid "hierarchy_id"
+    t.string "wizard_status"
     t.index ["hierarchy_type", "hierarchy_id"], name: "index_activities_on_hierarchy_type_and_hierarchy_id"
   end