+++ spec/factories/fund.rb
@@ -1,5 +1,6 @@
 FactoryBot.define do
   factory :fund do
     name { Faker::Company.name }
+    association :organisation, factory: :organisation
   end
 end